fixed some data declarations

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,8 +10,8 @@
  */
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int res;
-	int flip_count = 0;
+	unsigned long int res;
+	unsigned int flip_count = 0;
 
 	res = n ^ m;
 	while (res > 0)
